requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
configobj==5.0.6
configparser==5.0.1
cryptography==2.3
dbus-python==1.2.4
decorator==4.2.1
gpg==1.10.0
idna==2.5
isc==2.0
mtrpacket==1.0.0
netifaces==0.10.6
nftables==0.1
pciutils==2.3.6
perf==0.1
ply==3.9
pycairo==1.16.3
pycparser==2.14
pygobject==3.28.3
pyOpenSSL==18.0.0
pyparsing==2.1.10
python-dateutil==2.6.1
python-dmidecode==3.12.2
python-linux-procfs==0.6
pyudev==0.21.0
rhnlib==2.8.6
rpm==4.14.2
scapy==2.4.4
schedutils==0.6
selinux==2.9
sepolicy==1.1
setools==4.2.2
setroubleshoot==1.1
six==1.11.0
slip==0.6.4
slip.dbus==0.6.4
sos==3.8
speedtest-cli==2.1.2
SSSDConfig==2.2.3
syspurpose==1.26.20
systemd-python==234
